ファイルの仕分け、コードのリファクタリング
