+++ NewWillDiary/Diary/MakingCalendarViewController.swift
@@ -1,0 +1,20 @@
+//
+//  MakingCalendarViewController.swift
+//  NewWillDiary
+//
+//  Created by 澤田世那 on 2022/04/08.
+//
+
+import UIKit
+import FSCalendar
+import CalculateCalendarLogic
+
+class MakingCalendarViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        // Do any additional setup after loading the view.
+    }
+
+}
